pyproject.toml
[project]
name = "cozy-memory"
version = "0.2.0"
description = "统一记忆管理系统 - 整合 Memobase/Mem0/Cognee"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "蟹小五", email = "crab@example.com"}
]
keywords = ["memory", "memobase", "mem0", "cognee", "ai"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

memobase = ["memobase-client>=0.1.0"]
mem0 = ["mem0ai>=0.1.0"]
cognee = ["cognee>=0.1.0"]
redis = ["redis>=5.0.0"]

[project.scripts]
cozy-memory = "cozy_memory.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cozy_memory*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
python_files = "test_*.py"
addopts = "-v --cov=src --cov-report=html --cov-fail-under=80"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
